chore: suggested refactor
